+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### Version 1.3.8:
 - Added public function removeGroup(), which removes a group by name (and fires a new custom event).
+- Organized the public functions into more areas to make them easier to find.
 - 
 
 ### Version 1.3.7: